feat: diagnosa dto move
