+++ Model/Api/AddDiagnosaDto.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ProfiraClinic.Models.Api
+{
+    public class AddDiagnosaDto()
+    {
+        public string? KodeLokasi { get; set; }
+        public DateTime? TanggalTransaksi { get; set; }
+        public string? NomorAppointment { get; set; }
+        public string? KodeCustomer { get; set; }
+        public string? KodeKaryawan { get; set; }
+        public string? KodeDiagnosa { get; set; }
+        public string? KategoriDiagnosa { get; set; }
+        public string? KeteranganDiagnosa { get; set; }
+        public string? USRID { get; set; }
+        public string? NOFAK { get; set; }
+    }
+}
